ARC: Add check for defined but not implemented AUX registers
AUX regs are pretty special things as they are typically used for
controlling of certain HW features of ARC cores. And so typically
they are not like general purpose regs where we may write random values
and then read those values.

And so we'd better know early that there's a missing AUX reg as target
software may have some expectations on values of AUX regs.