[GDB] Add more auxiliary registers
build/config registers:
  timer_build
  irq_build
  vecbase_build
  isa_config
these registers can only be read and are not shown by default.

MMU registers:
  pid
  tlbpd0
  tlppd1
  tlbindex
  tlbcmd

note: the "field" values, e.g. VPN in tlbpd0, can be wrong. however,
the number reported as the value of the registers are fine. this is
probably a bug in GDB. I will investigate this further later.