[INST] Get rid of pre-allocated zero/one [2/2]
This patch ditches the whole ctx->one/ctx->zero.
As a result we can have translator_loop_temp_check()
at the end of each instruction translation and if
it prints a warning, then we do have a leakage issue.
Note that you must "configure" with "--enable-debug-tcg"
to see those warnings.

CAVEAT:
Yet again, the changes are done in generated code.
Issue #62 addresses this and asks for a fix in the
generator framework.

Signed-off-by: Shahab Vahedi <shahab@synopsys.com>