[MPU] Set the TLB page size correctly.
This is related to issue #93. There can be scenario's that a region
with higher priority is contained inside another. In such cases, we
have to be careful how we set the TLB's entry size. We set it to
TARGET_PAGE_SIZE only if it is not overlapping with another region
that has higher priority.

Signed-off-by: Shahab Vahedi <shahab@synopsys.com>