Making correct BTA implementation.
This patch implements BTA semantics as they are described in PRM.
This implementation fixes delayslot issues in all known cases.

It fixes the memcpy, strcpy issues that initially though would
be related to Zero Overhead Loops (ZOL) (aka Zero Delay Branches).

It also changes when blink is set in BL and JL instructions.
The blink is now set before the delayslot is executed.
PRM is ambiguous on this respect.